+++ lib-anim/src/main/java/com/heyanle/lib_anim/cycdm/CycdmParser.kt
@@ -4,10 +4,9 @@ import com.heyanle.lib_anim.*
 import com.heyanle.lib_anim.entity.Bangumi
 import com.heyanle.lib_anim.entity.BangumiDetail
 import com.heyanle.lib_anim.entity.BangumiSummary
-import com.heyanle.lib_anim.utils.OkHttpUtils
+import com.heyanle.lib_anim.utils.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import okhttp3.*
 import org.jsoup.Jsoup
 import com.google.gson.JsonParser
 import java.net.URLDecoder
@@ -259,39 +258,6 @@ class CycdmParser : ISourceParser, IHomeParser, IDetailParser, IPlayerParser, IS
         }
     }
 
-    private fun saltParseUrl (salt: String):String {
-        val client = OkHttpClient().newBuilder().build()
-
-        val body: RequestBody = MultipartBody.Builder().setType(MultipartBody.FORM)
-            .addFormDataPart("url", salt)
-            .addFormDataPart("time", "1674837556")
-            .addFormDataPart("key", "533fcd74ffa5c3fbb99a45c593d369fa")
-            .build()
-
-        val request: Request = Request.Builder()
-            .url("https://player.cycdm01.top/api_config.php")
-            .method("POST", body)
-            .addHeader("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/93.0.4577.82 Safari/537.36 Edg/93.0.961.52")
-            .addHeader("Accept", "*/*")
-            .addHeader("Host", "player.cycdm01.top")
-            .addHeader("Connection", "keep-alive")
-            .addHeader(
-                "Content-Type",
-                "multipart/form-data; boundary=--------------------------441235884103066496960501"
-            )
-            .build()
-
-        kotlin.runCatching {
-            val response = client.newCall(request).execute()
-            val jsonObject = JsonParser.parseString(response.body!!.string()).asJsonObject
-            return jsonObject.get("url").asString
-        }.getOrElse {
-            it.printStackTrace()
-        }
-
-        return salt
-    }
-
     override suspend fun getPlayUrl(
         bangumi: BangumiSummary,
         lineIndex: Int,
@@ -334,11 +300,23 @@ class CycdmParser : ISourceParser, IHomeParser, IDetailParser, IPlayerParser, IS
                 val playPattern = Regex("""(?<="url":").*(?=","u)""")
                 val playSecret = playPattern.find(playInfo)?.value ?: ""
 
-                var result = playSecret.base64Decoded
+                var result = Base64Utils.decode(playSecret)
                 result = URLDecoder.decode(result, "utf-8")
 
                 if (result.startsWith("cycdm")) {
-                    result = this@CycdmParser.saltParseUrl(result)
+                    kotlin.runCatching {
+                        val requestForUrl = OkHttpUtils.post(
+                            "https://player.cycdm01.top/api_config.php",
+                            OkHttpUtils.getPostFormBody()
+                                .addFormDataPart("url", result)
+                                .build()
+                        )
+
+                        val jsonObject = JsonParser.parseString(requestForUrl).asJsonObject
+                        result =  jsonObject.get("url").asString
+                    }.onFailure {
+                        it.printStackTrace()
+                    }
                 }
 
                 if (result.isNotEmpty())
